// src/app/model/shared/lojas.ts
export class Lojas {
       id_loj: number;
       titulo_loj_app: string;
       titulo_loj: string;
       latitude: number;
       longitude: number;
       rua_loj: string;
       bairro_loj: string;
       cidade_loj: string;
       estado_loj: string;
       numero_loj: string;
       id_loj_devolucao?: number;
       hora_retirada: any;
       id_loj_esta: any;
       nome_loj_esta: any;
}